Approve.
1. Add auth.
2. add data SensorData.
3. Add register.

+++ LibCore/NoAuthSession.cs
@@ -9,6 +9,6 @@ namespace Zcore.Dto
     public class NoAuthSession : IUserSession
     {
         public bool State { get; } = false;
-        public int UserId { get; } = Int32.MinValue;
+        public long UserId { get; } = long.MinValue;
     }
 }
